Pre-adjust some commands to run on all packages
